20210618 commit 출석 관련 파일 이동
