+++ shms/src/main/java/kr/ac/shms/lms/student/service/StudentService.java
@@ -39,9 +39,7 @@ import kr.ac.shms.main.commuity.vo.ComCodeVO;
  * </pre>
  */	
 public interface StudentService {
-
 	public void selectMainSuganList(Map<String,Object> paramMap);
-	
 	
 	/**
 	 * 한 학생에 대한 정보 추출
@@ -49,40 +47,6 @@ public interface StudentService {
 	 * @return 데이터가 없으면 null 반환
 	 */
 	public StudentVO student(String id);
-	
-	/**
-	 * 데이터 계산을 위해 정보 추출
-	 * @param studentInfo 학번, 강의코드
-	 * @return attendVO
-	 */
-	public AttendVO selectAttendInfo(AttendVO studentInfo);
-	
-	/**
-	 * 입실시간, 퇴실시간 추출
-	 * @param attendVO 학번, 강의 코드
-	 * @return 입실시간
-	 */
-	public AttendVO selectAtndanTime(AttendVO attendVO);
-	/**
-	 * qr코드에 넣을 정보 추출
-	 * @param attendInfo 학번, 강의코드
-	 * @return 학번, 강의코드
-	 */
-	
-	public AttendVO selectQRInfo(AttendVO attendInfo);
-	/**
-	 * 출석(입실)정보를 insert
-	 * @param attendInfo 학번, 강의코드
-	 * @return ServiceResult 
-	 */
-	public ServiceResult attend(AttendVO attendInfo);
-	
-	/**
-	 * 출석(퇴실)정보를 update
-	 * @param exitInfo 학번, 강의코드
-	 * @return ServiceResult
-	 */
-	public ServiceResult exit(AttendVO exitInfo);
 	
 	/**
 	 * 학생의 도서 대출에 대한 정보 추출
@@ -273,31 +237,10 @@ public interface StudentService {
      * @return 데이터가 없으면 null
      */
     public List<MypageVO> studentSubList(String sub_code);
-    
-    /**
-     * 입실 카운트
-     * @param attendVO
-     * @return
-     */
-    public ServiceResult selectCountAttend(AttendVO attendVO);
-    
-    /**
-     * 퇴실 카운트
-     * @param attendVO
-     * @return
-     */
-    public ServiceResult selectCountExit(AttendVO attendVO);
-    
+
     public List<StudentVO> passwordUpdate();
     
     public int passUpdate(UserLoginVO userLogin);
-
-    /**
-     * 출석상태 업데이트 
-     * @param attendVO
-     * @return 
-     */
-    public ServiceResult updateAttendStat(AttendVO attendVO);
     
     /**
      * 이력서/자소서 첨삭 리스트 출력
