+++ shms/src/main/resources/kr/ac/shms/db/mybatis/mappers/Attendance.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="kr.ac.shms.lms.student.dao.StudentDAO">
+<mapper namespace="kr.ac.shms.lecture.dao.LectureStudentDAO">
 	<select id="selectAtndanTime" resultType="AttendVO" parameterType="AttendVO">
 	SELECT attend_time
 		   , exit_time
